# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.DepartmentRequest:
    properties:
      dept_description:
        type: string
      dept_name:
        type: string
      id:
        type: string
    required:
    - dept_name
    type: object
...
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.TestCatalogRequest:
    properties:
      dept_id:
        type: string
      id:
        type: string
      sample_type:
        type: string
      test_code:
        type: string
      test_name:
        type: string
      test_price:
        type: number
      turnaround_time:
        type: integer
    required:
    - dept_id
    - test_name
    type: object
...
  github_com_jenish-brainztechs_go-backend_internal_core_domain.TestCatalog:
    properties:
      code:
        type: string
      departmentID:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      sampleType:
        type: string
      testPrice:
        format: float64
        type: number
      turnAroundTime:
        type: integer
      updatedAt:
        type: string
    type: object
...
  internal_adapter_handler_http.errorResponse:
    properties:
      messages:
        example:
        - Error message 1
        - ' Error message 2'
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
  internal_adapter_handler_http.response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Laboratory Information System API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: LIS Backend API
  version: "1.0"
paths:
...
  /admin/lab-test/create-catalog:
    post:
      consumes:
      - application/json
      description: Create a new test catalog entry (admin only)
      parameters:
      - description: Test catalog details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.TestCatalogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_core_domain.TestCatalog'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create test catalog
      tags:
      - Lab Tests
  /admin/lab-test/create-department:
    post:
      consumes:
      - application/json
      description: Create a new lab department (admin only)
      parameters:
      - description: Department details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.DepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_core_domain.Department'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create department
      tags:
      - Lab Tests
  /admin/lab-test/create-panel:
    post:
      consumes:
      - application/json
      description: Create a new test panel (admin only)
      parameters:
      - description: Panel details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.PanelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_core_domain.Panels'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create panel
      tags:
      - Lab Tests
  /admin/lab-test/create-panel-component:
    post:
      consumes:
      - application/json
      description: Associate a test catalog with a panel (admin only)
      parameters:
      - description: Panel component details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.PanelComponentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create panel component
      tags:
      - Lab Tests
  /admin/lab-test/create-reference:
    post:
      consumes:
      - application/json
      description: Create a new reference range for a test parameter (admin only)
      parameters:
      - description: Reference range details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReferenceRangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create reference range
      tags:
      - Lab Tests
  /admin/lab-test/create-test-parameter:
    post:
      consumes:
      - application/json
      description: Create a new test parameter for a test catalog (admin only)
      parameters:
      - description: Test parameter details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.TestParameterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create test parameter
      tags:
      - Lab Tests
  /admin/lab-test/update-catalog:
    patch:
      consumes:
      - application/json
      description: Update an existing test catalog (admin only)
      parameters:
      - description: Test catalog details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.TestCatalogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update test catalog
      tags:
      - Lab Tests
  /admin/lab-test/update-department:
    patch:
      consumes:
      - application/json
      description: Update an existing department (admin only)
      parameters:
      - description: Department details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.DepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update department
      tags:
      - Lab Tests
  /admin/lab-test/update-panel:
    patch:
      consumes:
      - application/json
      description: Update an existing panel (admin only)
      parameters:
      - description: Panel details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.PanelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update panel
      tags:
      - Lab Tests
  /admin/lab-test/update-reference:
    patch:
      consumes:
      - application/json
      description: Update an existing reference range (admin only)
      parameters:
      - description: Reference range details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReferenceRangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update reference range
      tags:
      - Lab Tests
  /admin/lab-test/update-test-parameter:
    patch:
      consumes:
      - application/json
      description: Update an existing test parameter (admin only)
      parameters:
      - description: Test parameter details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.TestParameterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update test parameter
      tags:
      - Lab Tests
...
  /lab-test/list-catalog/{id}:
    get:
      consumes:
      - application/json
      description: Get all test catalogs for a given department
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.TestCatalogResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: List test catalogs by department
      tags:
      - Lab Tests
  /lab-test/list-department:
    get:
      consumes:
      - application/json
      description: Get all lab departments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.DepartmentResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: List departments
      tags:
      - Lab Tests
  /lab-test/list-panel-catalog/{id}:
    get:
      consumes:
      - application/json
      description: Get all panel components for a given panel
      parameters:
      - description: Panel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.PanelComponentResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: List panel components by panel
      tags:
      - Lab Tests
  /lab-test/list-panel/{id}:
    get:
      consumes:
      - application/json
      description: Get all panels for a given department
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.PanelResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: List panels by department
      tags:
      - Lab Tests
  /lab-test/list-reference/{id}:
    get:
      consumes:
      - application/json
      description: Get all reference ranges for a given test parameter
      parameters:
      - description: Test Parameter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReferenceRangeResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: List reference ranges by parameter
      tags:
      - Lab Tests
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"